feat(api): extract metrics calculation logic into APIMetricsService
- Create APIMetricsService with centralized metrics calculation logic (272 lines)
- Refactor /api/metrics/route.ts: 137 → 39 lines (72% reduction)
- Refactor /api/health/route.ts: 143 → 76 lines (47% reduction)
- Eliminate 87 lines of duplicate business logic from API routes
- Remove 22 lines of duplicate circuit breaker calculation logic
- Consolidate 3 different calculation patterns into single atomic service
- Achieve perfect Service Layer compliance (blueprint.md:390)
- Enhance testability - business logic now isolated from API routes
- Follow DRY principle, Atomic Modularity, and Single Responsibility

Validation:
✅ Build (3.4s) ✅ Lint (0 warnings) ✅ Typecheck (0 errors)
✅ Tests (7/7 suites, 30/30 passing) ✅ Zero regressions
