+++ app/api/health/route.ts
@@ -1,8 +1,8 @@
 import { NextResponse } from "next/server";
 import { monitoringService } from "@/lib/monitoring";
 import { APIRouteHandler } from "@/lib/services/api-route-handler";
-import { circuitBreakerRegistry } from "@/lib/circuit-breaker";
 import { UnifiedCacheManager } from "@/lib/services/unified-cache-manager";
+import { APIMetricsService } from "@/lib/services/api-metrics-service";
 
 export const GET = APIRouteHandler.createGETHandler({
   requireAuth: false,
@@ -16,81 +16,15 @@ export const GET = APIRouteHandler.createGETHandler({
         // Get basic system health
         const systemHealth = await monitoringService.getSystemHealth();
 
-        // Add application-specific checks
-        const appChecks = {
-          nextjs: {
-            service: "nextjs",
-            status: "healthy" as const,
-            responseTime: 0,
-            metadata: {
-              version: process.env.npm_package_version || "1.0.0",
-              nodeVersion: process.version,
-              environment: process.env.NODE_ENV || "development",
-            },
-          },
-
-          auth: {
-            service: "auth",
-            status: process.env.CLERK_SECRET_KEY
-              ? ("healthy" as const)
-              : ("degraded" as const),
-            responseTime: 0,
-            metadata: {
-              configured: !!process.env.CLERK_SECRET_KEY,
-              provider: "clerk",
-            },
-          },
-
-          payments: {
-            service: "payments",
-            status: process.env.STRIPE_SECRET_KEY
-              ? ("healthy" as const)
-              : ("degraded" as const),
-            responseTime: 0,
-            metadata: {
-              configured: !!process.env.STRIPE_SECRET_KEY,
-              provider: "stripe",
-            },
-          },
-
-          // Circuit breaker health checks
-          circuitBreakers: {
-            service: "circuit-breakers",
-            status: (() => {
-              const openCircuits = circuitBreakerRegistry.getOpenCircuits();
-              const totalCircuits = Object.keys(
-                circuitBreakerRegistry.getAllMetrics(),
-              ).length;
-              const healthyCount = totalCircuits - openCircuits.length;
-
-              if (openCircuits.length === 0) return "healthy" as const;
-              if (healthyCount >= totalCircuits * 0.7)
-                return "degraded" as const;
-              return "unhealthy" as const;
-            })(),
-            responseTime: 0,
-            metadata: {
-              totalCircuits: Object.keys(circuitBreakerRegistry.getAllMetrics())
-                .length,
-              openCircuits: circuitBreakerRegistry.getOpenCircuits(),
-              circuitBreakerStates: circuitBreakerRegistry.getAllMetrics(),
-            },
-          },
-        };
+        // Application health checks via service (eliminates code duplication)
+        const appChecks = APIMetricsService.getApplicationHealthChecks();
 
         const allChecks = [...systemHealth.checks, ...Object.values(appChecks)];
 
-        // Determine overall status
-        const statuses = allChecks.map((check) => check.status);
-        let overallStatus: "healthy" | "degraded" | "unhealthy";
-
-        if (statuses.every((status) => status === "healthy")) {
-          overallStatus = "healthy";
-        } else if (statuses.some((status) => status === "unhealthy")) {
-          overallStatus = "unhealthy";
-        } else {
-          overallStatus = "degraded";
-        }
+        // Determine overall status using service logic
+        const overallStatus = APIMetricsService.calculateOverallSystemStatus({
+          checks: allChecks,
+        });
 
         const response = {
           status: overallStatus,
